Исправлен баг в заполнении границ символов. Модулятор добавляет опорный символ, если он не включен в последовательность бит

+++ cpp-DPSK-modulation/dpsk_demodulator.h
@@ -22,7 +22,7 @@ namespace dpsk_demod {
         double ExtractPhaseValue(std::complex<double> inphase_quadrature_components) const;
 
         /// Демодуляция последовательности отсчетов. Сложность: O(???)
-        std::vector<uint16_t> Demodulation(const std::vector<double>& samples);
+        std::vector<uint32_t> Demodulation(const std::vector<double>& samples);
 
 //    private:
         /// Сгенерировать один период косинуса и синуса при заданных параметрах. Сложность(sampling_frequency / carrier_frequency)
@@ -32,10 +32,10 @@ namespace dpsk_demod {
         const std::vector<double>& GetBoundsSymbols() const noexcept;
 
         /// Получить последовательность символов на окружности в соответствии с кодом Грея. Сложность: O(1)
-        const std::vector<uint16_t>& GetSymbolsSequenceOnCircle() const noexcept;
+        const std::vector<uint32_t>& GetSymbolsSequenceOnCircle() const noexcept;
 
         /// Определить символ на основании значения разности фаз между символами. Сложность: O(positionality)
-        uint16_t DefineSymbol(double phase_difference) const noexcept;
+        uint32_t DefineSymbol(double phase_difference) const noexcept;
 
     private:
         /// Заполнить границы (сектора) символов на огружности. Сложность: O(positionality)
@@ -49,7 +49,7 @@ namespace dpsk_demod {
         std::vector<double> cos_oscillation_;
         std::vector<double> sin_oscillation_;
         std::vector<double> bounds_symbols_; // границы диапазонов разностей фаз между символами
-        std::vector<uint16_t> symbols_sequence_on_circle_; // полследовательность символов на окружности в соответствии с кодом Грея
+        std::vector<uint32_t> symbols_sequence_on_circle_; // полследовательность символов на окружности в соответствии с кодом Грея
     };
 
     namespace tests {
